[Refactor] 커뮤니티 페이지 컴포넌트 분리
